# Only copy the relevant portion of the screen when copying to backbuffer in Compatibility backend