Partially addresses: https://github.com/godotengine/godot/issues/79439

Implements the first point from https://github.com/godotengine/godot/issues/79439#issuecomment-1756481658

Previously we copied the entire backbuffer, now we only copy the relevant portion. This matches the behavior on the RD backends. 

Opening as draft as I need to do more testing