@Calinou 2023-11-24
Tested locally (rebased on top of `master` 671c04f89f449f94e9d129d2fa0b96d3f65dd191), it works as expected. Visuals look identical in the MRP.

## Exported projects (for quick testing)

*Press Enter/Space, left-click or tap the screen to toggle Clip Children.*

### Android

- [ClipChildrenSlow.apk.zip](https://github.com/godotengine/godot/files/13461396/ClipChildrenSlow.apk.zip)
- [ClipChildrenSlow.GL-CanvasGroup-Performance.apk.zip](https://github.com/godotengine/godot/files/13461397/ClipChildrenSlow.GL-CanvasGroup-Performance.apk.zip)

### Web 

- [ClipChildrenSlow.GL-CanvasGroup-Performance.web.zip](https://github.com/godotengine/godot/files/13461399/ClipChildrenSlow.GL-CanvasGroup-Performance.web.zip)
- [ClipChildrenSlow.web.zip](https://github.com/godotengine/godot/files/13461400/ClipChildrenSlow.web.zip)

## Benchmark

### Linux

**OS:** Fedora 39
**CPU:** Intel Core i9-13900K
**GPU:** GeForce RTX 4090 (NVIDIA 535.129.03)

#### 1152×648 window

No Clip Children | Clip Children Before | Clip Children After *(this PR)*
-|-|-
11594 FPS (0.09 mspf) | 859 FPS (1.16 mspf) | 942 FPS (1.06 mspf)

#### 3840×2160 fullscreen

No Clip Children | Clip Children Before | Clip Children After *(this PR)*
-|-|-
6611 FPS (0.15 mspf) | 393 FPS (2.54 mspf) | 882 FPS (1.13 mspf)

### Web on Linux

**Browser:** Chromium 118
**CPU:** Intel Core i9-13900K
**GPU:** GeForce RTX 4090 (NVIDIA 535.129.03)

*Start Chromium with the `--disable-frame-rate-limit --disable-gpu-vsync` CLI arguments to disable any FPS limiting from the browser.*

#### 3840×2160 fullscreen

No Clip Children | Clip Children Before | Clip Children After *(this PR)*
-|-|-
2431 FPS (0.41 mspf) | 282 FPS (3.55 mspf) | 447 FPS (2.24 mspf)

### Android

**OS:** Android 13
**Device:** Samsung Galaxy Z Fold4 (Snapdragon 8 Gen 1+, Adreno 730)

*Measured using the main display, which has a resolution of 2176×1812.*

No Clip Children | Clip Children Before | Clip Children After *(this PR)*
-|-|-
120 FPS (8.33 mspf, capped by V-Sync) | 62 FPS (16.13 mspf) | 120 FPS (8.33 mspf, capped by V-Sync)

### Web on Android

**Browser:** Bromite 118
**Device:** Samsung Galaxy Z Fold4 (Snapdragon 8 Gen 1+, Adreno 730)

*Measured using the main display, which has a resolution of 2176×1812.*

No Clip Children | Clip Children Before | Clip Children After *(this PR)*
-|-|-
120 FPS (8.33 mspf, capped by V-Sync) | 10 FPS (100.00 mspf) | 13 FPS (76.92 mspf)