@clayjohn 2024-01-06
Thank you @Calinou for testing I wanted to ensure that a performance difference could be reproduced on Android, since you already did that, I am happy to mark as ready for review