@YuriSizov 2024-01-24
Cherry-picked for 4.2.2.