@Nicholas3413 2024-09-02
Are there anyway to include this improvement for the Godot 4.1.5 next? i tested the FPS difference on Godot 4.3 (stable) and Godot 4.1.4, there are about 15~ fps gap  (145 vs 131)