@clayjohn 2024-09-02
> Are there anyway to include this improvement for the Godot 4.1.5 next? i tested the FPS difference on Godot 4.3 (stable) and Godot 4.1.4, there are about 15~ fps gap (145 vs 131)

I'm not sure if we will be releasing another 4.1.x version. 

However, in either case, a new PR would be needed targeting 4.1.x directly as the source diverged enough between 4.1 and 4.2 that this can't be cleanly cherry picked. In other words, you will have to copy these changes manually into the 4.1 branch if you want to use them there.